Refactor into modules and add argument parser
