+++ README.md
@@ -1,6 +1,8 @@
 # autogitpull
-Automatic Git Puller &amp; Monitor
+Automatic Git Puller & Monitor
 
 By default, repositories whose `origin` remote does not point to GitHub or require
 authentication are skipped during scanning. Pass `--include-private` after the
 root folder to include these repositories.
+
+Use `--show-skipped` to display repositories that were skipped in the TUI.
