+++ compile-cl.bat
@@ -1,1 +1,1 @@
-cl /std:c++17 /EHsc git_auto_pull_all.cpp
+cl /std:c++17 /EHsc autogitpull.cpp git_utils.cpp tui.cpp arg_parser.cpp
