+++ compile.bat
@@ -1,1 +1,1 @@
- g++ -std=c++17 -o git_auto_pull_all.exe autogitpull.cpp
+ g++ -std=c++17 -o git_auto_pull_all.exe autogitpull.cpp git_utils.cpp tui.cpp arg_parser.cpp
