+++ git_utils.h
@@ -1,0 +1,22 @@
+#ifndef GIT_UTILS_H
+#define GIT_UTILS_H
+
+#include <filesystem>
+#include <string>
+
+namespace git {
+namespace fs = std::filesystem;
+
+std::string run_cmd(const std::string& cmd, int timeout_sec = 0);
+bool is_git_repo(const fs::path& p);
+std::string get_local_hash(const fs::path& repo);
+std::string get_current_branch(const fs::path& repo);
+std::string get_remote_hash(const fs::path& repo, const std::string& branch);
+std::string get_origin_url(const fs::path& repo);
+bool is_github_url(const std::string& url);
+bool remote_accessible(const fs::path& repo);
+int try_pull(const fs::path& repo, std::string& out_pull_log);
+
+} // namespace git
+
+#endif // GIT_UTILS_H
